+++ gulpfile.js
@@ -1,4 +1,4 @@
-/*!
+/**
  * @author:    Divio AG
  * @copyright: http://www.divio.ch
  */
@@ -7,18 +7,18 @@
 
 // #############################################################################
 // #IMPORTS#
-var autoprefixer = require('gulp-autoprefixer');
 var gulp = require('gulp');
 var gutil = require('gulp-util');
+var autoprefixer = require('autoprefixer');
+var postcss = require('gulp-postcss');
 var gulpif = require('gulp-if');
-var karma = require('karma').server;
 var sass = require('gulp-sass');
 var iconfont = require('gulp-iconfont');
 var iconfontCss = require('gulp-iconfont-css');
 var sourcemaps = require('gulp-sourcemaps');
 var minifyCss = require('gulp-minify-css');
-var protractor = require('gulp-protractor').protractor;
-var webdriverUpdate = require('gulp-protractor').webdriver_update;
+var eslint = require('gulp-eslint');
+var integrationTests = require('djangocms-casper-helpers/gulp');
 
 var argv = require('minimist')(process.argv.slice(2));
 
@@ -29,20 +29,35 @@ var options = {
 };
 var PROJECT_ROOT = __dirname;
 var PROJECT_PATH = {
-    'sass': PROJECT_ROOT + '/djangocms_admin_style/sass',
-    'css': PROJECT_ROOT + '/djangocms_admin_style/static/djangocms_admin_style/css',
-    'tests': PROJECT_ROOT + '/tests',
-    'icons': PROJECT_ROOT + '/djangocms_admin_style/static/djangocms_admin_style/fonts'
+    sass: PROJECT_ROOT + '/djangocms_admin_style/sass',
+    css: PROJECT_ROOT + '/djangocms_admin_style/static/djangocms_admin_style/css',
+    js: PROJECT_ROOT + '/djangocms_admin_style/static/djangocms_admin_style/js',
+    tests: PROJECT_ROOT + '/djangocms_admin_style/tests',
+    icons: PROJECT_ROOT + '/djangocms_admin_style/static/djangocms_admin_style/fonts'
 };
 
 var PROJECT_PATTERNS = {
-    'sass': [
+    sass: [
         PROJECT_PATH.sass + '/**/*.{scss,sass}'
     ],
     icons: [
         PROJECT_PATH.icons + '/src/*.svg'
+    ],
+    js: [
+        PROJECT_PATH.js + '/**/*.js',
+        PROJECT_PATH.tests + '/**/*.js',
+        '!' + PROJECT_PATH.js + '/**/jquery.*.js',
+        'gulpfile.js'
     ]
 };
+
+var INTEGRATION_TESTS = [
+    [
+        'loginAdmin',
+        'dashboard',
+        'addNewUser'
+    ]
+];
 
 // #############################################################################
 // #TASKS#
@@ -53,10 +68,11 @@ gulp.task('sass', function () {
         .on('error', function (error) {
             gutil.log(gutil.colors.red('Error (' + error.plugin + '): ' + error.messageFormatted));
         })
-        .pipe(autoprefixer({
-            // browsers are coming from browserslist file
-            cascade: false
-        }))
+        .pipe(postcss([
+            autoprefixer({
+                cascade: false
+            })
+        ]))
         .pipe(minifyCss({
             rebase: false
         }))
@@ -76,48 +92,41 @@ gulp.task('icons', function () {
         fontName: 'django-admin-iconfont',
         normalize: true
     }))
-    .on('glyphs', function (glyphs, options) {
-        gutil.log.bind(glyphs, options);
+    .on('glyphs', function (glyphs, opts) {
+        gutil.log.bind(glyphs, opts);
     })
     .pipe(gulp.dest(PROJECT_PATH.icons));
 });
 
+
+gulp.task('lint', ['lint:javascript']);
+gulp.task('lint:javascript', function () {
+    // DOCS: http://eslint.org
+    return gulp.src(PROJECT_PATTERNS.js)
+        .pipe(eslint())
+        .pipe(eslint.format())
+        .pipe(eslint.failAfterError());
+});
+
 // #######################################
 // #TESTS#
-gulp.task('tests', ['tests:unit', 'tests:integration']);
-gulp.task('tests:unit', function (done) {
-    // run javascript tests
-    karma.start({
-        configFile: PROJECT_PATH.tests + '/karma.conf.js',
-        singleRun: true
-    }, done);
-});
+gulp.task('tests', ['tests:integration']);
 
-gulp.task('tests:webdriver', webdriverUpdate);
-gulp.task('tests:integration', ['tests:webdriver'], function () {
-    return gulp.src([PROJECT_PATH.tests + '/integration/specs/*.js'])
-        .pipe(protractor({
-            configFile: PROJECT_PATH.tests + '/protractor.conf.js',
-            args: []
-        }))
-        .on('error', function (error) {
-            gutil.log(gutil.colors.red(
-                'Error (' + error.plugin + '): ' + error.message)
-            );
-        });
-});
-
-gulp.task('tests:watch', function () {
-    // run javascript tests
-    karma.start({
-        configFile: PROJECT_PATH.tests + '/karma.conf.js'
-    });
-});
+// gulp tests:integration [--clean] [--screenshots] [--tests=loginAdmin,toolbar]
+gulp.task('tests:integration', integrationTests({
+    tests: INTEGRATION_TESTS,
+    pathToTests: PROJECT_PATH.tests,
+    argv: argv,
+    dbPath: 'testdb.sqlite',
+    serverCommand: 'testserver.py',
+    logger: gutil.log.bind(gutil)
+}));
 
 // #############################################################################
 // #COMMANDS#
 gulp.task('watch', function () {
     gulp.watch(PROJECT_PATTERNS.sass, ['sass']);
+    gulp.watch(PROJECT_PATTERNS.js, ['lint']);
 });
 
-gulp.task('default', ['sass', 'watch']);
+gulp.task('default', ['sass', 'lint', 'watch']);